fix crash when resuming message propagation with MaxMessages
the problem is triggered by the source-side message disappearing
after a transaction to propagate it was started and then interrupted.

it seems tempting to centralize the null-check, but some of the other
branches are taken in situations where the relevant messages from the
source store have not been requested.

no autotest, as our test suite does not support injecting changes before
resuming.

amends 0089f49.