do not let both-sided uidvalidity change deter us
the algorithm is symmetrical, comparing the msgids that belong to the
paired uids. so it doesn't matter for recovering one side if the other
side's uidvalidity also changed. it does however impact our ability to
say on which side the change was genuine.

the pointless limitation was presumably a vestige from an earlier
iteration.

amends 77acc26 and 594e60b.