stub out ChangeLog
that file is a relic from the times before distributed VCS with atomic
commits existed. ours was slightly filtered, but still way too noisy to
be useful for end users. so basically, it had no target audience.

the file itself is left in place, because automake complains otherwise
(we don't want to use 'foreign' strictness, as otherwise INSTALL is not
added by autoreconf --install).